chore: move classifiers to own file
