+++ packages/playground/src/api/documents/shared.ts
@@ -73,27 +73,6 @@ export interface DocumentsSourceFileFilter {
   size?: Range<number>;
 }
 
-export interface ClassifierName {
-  name?: string;
-}
-
-export interface Classifier extends ClassifierName {
-  id: CogniteInternalId;
-  projectId?: number;
-  createdAt?: number;
-  status?: string;
-  active?: boolean;
-  metrics?: ClassifierMetrics;
-}
-
-export interface ClassifierMetrics {
-  precision?: number;
-  recall?: number;
-  f1Score?: number;
-  confusionMatrix?: number[][];
-  labels?: string[];
-}
-
 export interface ExternalDocumentsSearch {
   filter?: DocumentsFilter;
   search?: DocumentsSearch;
